Frequency files with Borger numbers
Finally, frequency lists for NA, NB, OB, and Ebla with Borger numbers for the values that fall within our 192 signs.
They are generated with different sets of code depending on the data (one file or thousands of files), etc. but looking up Borger numbers is universal, and we can generalize more of it later, but it is not based on live data anyway, so it made sense to do all kinds of different things to clean up the data a little ad-hocly.